implementation/requirements.txt
pip
uwsgi
psycopg2-binary
django
filetype
pillow
requests
django-colorfield
django-ckeditor
django-cors-headers
django-archive
django-import-export
django-registration
django-registration-redux
django-settings-export
django-forms-bootstrap
django-crispy-forms
django-rosetta
django-template-check
django-dia
django-pycdi
django-page-components
django-method-override
django-leaflet
unicode-slugify
pylint-django
beautifulsoup4
html2text
html5lib
gdal
ics
